juriBot/requirements.txt
fastapi
uvicorn
pydantic
requests
mistralai
python-dotenv
requests